✨(lib-components) serviceWorkerRefreshToken utility
Create a partial service-worker `serviceWorkerRefreshToken` in lib-components.
It will manage the refresh token after a 401.

Create a hook useServiceWorkerRefreshToken to communicate
 with this service worker.
This hook communicates with the service worker thanks to 3 actions:
- `getRefreshToken`: communicate the refresh token to the
 service worker when asked
- `setToken`: refresh the JWT token when it got the instruction
 from the service worker
- `logout`: clear the jwt and so logout when it got the instruction
 from the service worker
