+++ src/frontend/apps/standalone_site/src/features/Authentication/components/Authenticator.tsx
@@ -1,4 +1,10 @@
-import { AnonymousUser, Loader, useCurrentUser, useJwt } from 'lib-components';
+import {
+  AnonymousUser,
+  Loader,
+  useCurrentUser,
+  useJwt,
+  useServiceWorkerRefreshToken,
+} from 'lib-components';
 import { Fragment, PropsWithChildren, useEffect } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
@@ -6,10 +12,6 @@ import { routes } from 'routes';
 
 import { getCurrentUser } from '../api/getUserData';
 import { validateChallenge } from '../api/validateChallenge';
-import {
-  EServiceworkerAuthAction,
-  ServiceworkerAuthMessage,
-} from '../model/service-worker';
 
 const TARGET_URL_STORAGE_KEY = 'redirect_uri';
 const QUERY_PARAMS_CHALLENGE_TOKEN_NAME = 'token';
@@ -24,7 +26,8 @@ export const Authenticator = ({ children }: PropsWithChildren<unknown>) => {
     currentUser: state.currentUser,
     setCurrentUser: state.setCurrentUser,
   }));
-  const { jwt, setJwt, resetJwt, refreshJwt, setRefreshJwt } = useJwt();
+  const { jwt, setJwt, resetJwt } = useJwt();
+  useServiceWorkerRefreshToken();
 
   useEffect(() => {
     if (jwt) {
@@ -66,63 +69,6 @@ export const Authenticator = ({ children }: PropsWithChildren<unknown>) => {
     };
     fetchUserData();
   }, [currentUser, history, jwt, pathname, resetJwt, setCurrentUser]);
-
-  /**
-   * This effect is used to communicate with the service worker
-   */
-  useEffect(() => {
-    if (!('serviceWorker' in navigator)) {
-      return;
-    }
-
-    /**
-     * This handle communicates with the service worker, 3 actions are possible:
-     *  - `getRefreshToken`: communicate the refresh token to the service worker when asked
-     *  - `setToken`: refresh the JWT token when it got the instruction from the service worker
-     *  - `logout`: clear the jwt and so logout when it got the instruction from the service worker
-     */
-    const handleFetch = (event: MessageEvent) => {
-      const { action, requestId, valueSW } =
-        event.data as ServiceworkerAuthMessage;
-
-      switch (action) {
-        case EServiceworkerAuthAction.GET_ACCESS_TOKEN:
-          if (event.source) {
-            event.source.postMessage({
-              action: EServiceworkerAuthAction.ACCESS_TOKEN_RESPONSE,
-              valueClient: jwt,
-              requestId: requestId,
-            });
-          }
-          break;
-        case EServiceworkerAuthAction.GET_REFRESH_TOKEN:
-          if (event.source) {
-            event.source.postMessage({
-              action: EServiceworkerAuthAction.REFRESH_TOKEN_RESPONSE,
-              valueClient: refreshJwt,
-              requestId: requestId,
-            });
-          }
-          break;
-        case EServiceworkerAuthAction.SET_TOKEN:
-          if (valueSW) {
-            setJwt(valueSW.access);
-            setRefreshJwt(valueSW.refresh);
-          }
-          break;
-        case EServiceworkerAuthAction.LOGOUT:
-        default:
-          resetJwt();
-          break;
-      }
-    };
-
-    navigator.serviceWorker.addEventListener('message', handleFetch);
-
-    return () => {
-      navigator.serviceWorker.removeEventListener('message', handleFetch);
-    };
-  }, [refreshJwt, setJwt, setRefreshJwt, resetJwt, jwt]);
 
   if (!jwt && !code) {
     return <Fragment />;
