+++ src/frontend/packages/lib_components/src/hooks/stores/index.ts
@@ -1,5 +1,5 @@
 export * from './useCurrentResourceContext';
+export * from './useCurrentSession';
 export * from './useCurrentUser';
 export * from './useJwt';
 export * from './useMaintenance';
-export * from './useCurrentSession';
