+++ src/frontend/packages/lib_components/src/types/index.ts
@@ -1,15 +1,15 @@
+export * from './AppData';
 export * from './api';
-export * from './User';
-export * from './jwt';
-export * from './ResourceContext';
 export * from './apps/';
-export * from './tracks';
-export * from './api';
-export * from './models';
-export * from './XMPP';
 export * from './file';
+export * from './jwt';
+export * from './models';
 export * from './Participant';
 export * from './PortabilityRequest';
 export * from './Organization';
-export * from './AppData';
+export * from './ResourceContext';
+export * from './serviceWorker';
+export * from './tracks';
+export * from './User';
+export * from './XMPP';
 export * from './XAPI';
