+++ src/frontend/packages/lib_components/src/types/serviceWorker.ts
@@ -1,4 +1,4 @@
-import { TokenResponse } from 'lib-components';
+import { TokenResponse } from './jwt';
 
 export enum EServiceworkerAuthAction {
   GET_REFRESH_TOKEN = 'GET_REFRESH_TOKEN',
