+++ src/frontend/packages/lib_components/src/utils/serviceWorker/index.ts
@@ -1,1 +1,2 @@
 export * as serviceWorkerRegistration from './serviceWorkerRegistration';
+export * from './serviceWorkerRefreshToken';
