+++ src/frontend/packages/lib_components/src/utils/serviceWorker/serviceWorkerRefreshToken/index.ts
@@ -1,0 +1,2 @@
+export * from './useServiceWorkerRefreshToken';
+export * as serviceWorkerRefreshToken from './serviceWorkerRefreshToken';
